01-instrument-fastapi/app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
prometheus-client==0.21.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-requests==0.49b2
structlog==24.4.0
langfuse
pyroscope-io
